+++ apps/nextjs-website/src/app/solutions/page.tsx
@@ -1,11 +1,11 @@
-import SolutionsTemplate from '@/components/templates/SolutionsTemplate/SolutionsTemplate';
+import SolutionListTemplate from '@/components/templates/SolutionListTemplate/SolutionListTemplate';
 import { getSolutionsList } from '@/lib/api';
 
 const Page = async () => {
   const solutionsList = await getSolutionsList();
 
   return (
-    <SolutionsTemplate
+    <SolutionListTemplate
       hero={solutionsList.hero}
       solutions={solutionsList.solutions}
       features={solutionsList.features}